improve 2024 day 6 part 2
My first approach was to brute force all the possible places
a block could go, but discussion on Reddit made me understand
that the only places a new block could go would be on places
that the guard might be.

So, first run the algorithm until the guard leaves the area,
and then test whether adding a block at each step along the way
would produce a loop. This reduces the number of possible block
locations by over half.